system administrator implementation

+++ src/main/java/com/example/socapplication/repository/AppUserRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AppUserRepository extends ListCrudRepository<AppUser, Integer> {
+public interface AppUserRepository extends ListCrudRepository<AppUser, Long> {
     Optional<AppUser> findByEmail(String email);
 }
